Update to C++20, now it's 2026. Fix matters arising.
* fix some u8 literal stuff, including updating IconFontCppHeaders (see
  https://github.com/juliettef/IconFontCppHeaders/issues/15)

* remove some variables that were always unused, but somehow not
  previously a problem

* replace ill-advised volatile++ with probably equally ill-advised
  volatile=volatile+1

* tweak BeebLink code due to aggregate initialization changes

* add casts to test_AVFoundation code

+++ experimental/test_AVFoundation/test_AVFoundation.mm
@@ -440,9 +440,9 @@ int main(int argc, char *argv[]) {
             asbd.mBitsPerChannel = fmt->wBitsPerSample;
             asbd.mBytesPerFrame = asbd.mChannelsPerFrame * asbd.mBitsPerChannel / 8;
             asbd.mBytesPerPacket = asbd.mFramesPerPacket * asbd.mBytesPerFrame;
-            asbd.mFormatFlags = (kAudioFormatFlagIsSignedInteger |
-                                 kAudioFormatFlagsNativeEndian |
-                                 kAudioFormatFlagIsPacked);
+            asbd.mFormatFlags = ((UInt32)kAudioFormatFlagIsSignedInteger |
+                                 (UInt32)kAudioFormatFlagsNativeEndian |
+                                 (UInt32)kAudioFormatFlagIsPacked);
 
             CMFormatDescriptionRef afd_ref;
             OSStatus result = CMAudioFormatDescriptionCreate(kCFAllocatorDefault, &asbd, 0, nullptr, 0, nullptr, nullptr, &afd_ref);
